Significantly speed up content database writes.
Content database data loss is tolerable, so the fsync can be skipped.
Unlike tags, it's not personal user data (the content came from a web
server), it's likely to be rediscovered on the next feed update anyway,
and Elfeed can operate normally with missing content. It just can't
display that entry with elfeed-show-mode.

This increases the write speed of the content database by about 100x, a
significant gain at virtually no cost.