Introduced hooks for http and parse errors
I used Xah's advice on http://ergoemacs.org/emacs_manual/elisp/Hooks.html and end the name of the hooks with "hooks" instead of "hook" as it is an "abnormal hook" because the hooks are called with arguments. He says: "By convention, abnormal hook names end in ‘-functions’ or ‘-hooks’. If the variable's name ends in ‘-function’, then its value is just a single function, not a list of functions.". I also made sure not to change the return value of the "handle" functions by keeping the message call as the last called function.