Index upgrade routine for Emacs 26 (fixes #207, #229)
Emacs 26 introduces a new "record" type, and defstruct objects are now
based on this type. Unfortunately this changes (and breaks) Elfeed's
index format. Prior to this commit, Emacs 26 could not open an Emacs 25
index and vice versa. After this change, Elfeed running under Emacs 26
will automatically and quietly upgrade an Emacs 25 database index for
Emacs 26. THIS UPGRADE IS IRREVERSIBLE, but a backup copy of the Emacs
25 index is made before this upgrade is performed.

To keep Emacs 25 from hitting a nasty reader error attempting to load
the new database format, an old-style dummy database is now inserted
before the real database. Elfeed running under Emacs 25 will only load
this empty database. This means the contents will be lost on the next
index save, but that's what the backup file is all about.

In general, the backup file can be safely deleted if the upgrade
completed successfully and the user no longer plans to run Elfeed under
Emacs 25 and use this database.

The previous database format version was "0.0.3" (a string) and this is
what Emacs 25 will continue to use. The new database format version is a
integer 4. Format versions will be integers from here on.

The previous upgrade routine has been discarded since it's assumed all
existing databases with any value have already been upgraded to the
previous version.