Dynamically enable/disable --compressed curl option
It's possible to build curl without compression support (libz), and
indeed there are some binaries out there without it. Just like with
version sniffing, Elfeed now detects if the installed version of curl
supports compression by sniffing for "libz" from --version. This is
essentially "free" since it's already inspecting the version number.

The elfeed-curl-get-version function has been replaced with a new
elfeed-curl-get-capabilities that returns a richer plist of information.
Currently it's just a version string and boolean for compression
support.

This fixes #276.