Allow fetching over HTTP/2
I had to disable HTTP/2 back in June 2016 when I first implemented the
curl backend because a lot of hosts didn't implement the protocol
correctly. Feeds that fetched fine with HTTP/1.1 had bizarre errors when
fetched with HTTP/2. The simplest solution was just to disable HTTP/2.

In the past 2.5 years, it seems all those bugs have been ironed out, so
it's time to allow Elfeed to use HTTP/2.