Add :last-seen metadata to all entries (WIP, #357)
This change is incomplete due to 304 Not Modified responses. If the feed
hasn't been modified, the entries aren't actually seen, and so they do
not have their :last-seen tags updated. However, this doesn't accurately
reflect the intention of the :last-seen slot.

To fix this, in addition to :etag and :last-modified, Elfeed would need
to keep a list of the entries seen in that update so that it can update
their :last-seen tags — as if virtually fetching the feed again.