Added elfeed-entry-other-window function
In elfeed-search mode, this function opens the elfeed entry under point in
another window if another window is present, or else sensibly splits the frame
if there is only a single window and opens the elfeed entry in the other
window. After opening the elfeed entry in the other window, the pointer is
moved down one line in elfeed search.

Can be bound to "o" by adding the following to your dot emacs file:
(add-hook 'elfeed-search-mode-hook (lambda () (local-set-key (kbd "o") 'elfeed-entry-other-window)))