Refactor entry rendering after eww and optionally to retrieve content
Refactor entry rendering after eww. Now there's a separate render
function that can specifically handle each content type.
The render function is used as callback for eww if the retrieve
of content through the entry link is enabled.
If this is not used the render function is called directly with the
content string.

The retrieval of content option allows for several conditions to
enable it such as always, regular expression match against the article
link or length of content.